requirements.txt
undetected_chromedriver==3.5.5
selenium==4.26.1
python-fasthtml==0.9.1